A lot of progress on typing agent.py
