+++ src/blop/agent.py
@@ -3,11 +3,11 @@ import os
 import pathlib
 import time as ttime
 import warnings
-from abc import ABC, abstractmethod
 from collections import OrderedDict
-from collections.abc import Mapping, Sequence
-from typing import Callable, Optional, Union
+from collections.abc import Mapping
+from typing import Any, Callable, Optional, Sequence, Union, Iterator, Generator
 
+from bluesky.run_engine import Msg
 import bluesky.plan_stubs as bps  # noqa F401
 import bluesky.plans as bp  # noqa F401
 import botorch  # type: ignore[import-untyped]
@@ -22,6 +22,7 @@ from botorch.acquisition.objective import ScalarizedPosteriorTransform  # type: 
 from botorch.models.deterministic import GenericDeterministicModel  # type: ignore[import-untyped]
 from botorch.models.model_list_gp_regression import ModelListGP  # type: ignore[import-untyped]
 from botorch.models.transforms.input import Normalize  # type: ignore[import-untyped]
+from botorch.models.model import Model  # type: ignore[import-untyped]
 from databroker import Broker  # type: ignore[import-untyped]
 from numpy.typing import ArrayLike
 from ophyd import Signal  # type: ignore[import-untyped]
@@ -63,32 +64,13 @@ def _validate_dofs_and_objs(dofs: DOFList, objs: ObjectiveList):
                     )
 
 
-class BlueskyAdaptiveBaseAgent(ABC):
-    """Placeholder for inheritance for clarity, while bluesky-adaptive deps is a moving target.
-    E.g. some versions of databroker v2 will break the current implementation.
-    """
-
-    @abstractmethod
-    def measurement_plan(self, point: ArrayLike) -> Tuple[str, List, dict]: ...
-
-    @staticmethod
-    @abstractmethod
-    def unpack_run(run) -> Tuple[Union[float, ArrayLike], Union[float, ArrayLike]]: ...
-
-    @abstractmethod
-    def tell(self, x, y) -> Dict[str, ArrayLike]: ...
-
-    @abstractmethod
-    def ask(self, batch_size: int) -> Tuple[Sequence[Dict[str, ArrayLike]], Sequence[ArrayLike]]: ...
-
-
 class BaseAgent:
     def __init__(
         self,
         *,
         dofs: Sequence[DOF],
         objectives: Sequence[Objective],
-        acquistion_plan: Optional[Union[Callable, str]] = default_acquisition_plan,
+        acquisition_plan: Callable = default_acquisition_plan,
         digestion: Callable = default_digestion_function,
         digestion_kwargs: Optional[dict] = None,
         verbose: bool = False,
@@ -97,6 +79,7 @@ class BaseAgent:
         model_inactive_objectives: bool = False,
         tolerate_acquisition_errors: bool = False,
         sample_center_on_init: bool = False,
+        train_every: int = 1,
     ):
         """_summary_
 
@@ -106,7 +89,7 @@ class BaseAgent:
             The degrees of freedom that the agent can control, which determine the output of the model.
         objectives : Sequence[Objective]
             The objectives which the agent will try to optimize.
-        acquistion_plan : Callable, optional
+        acquisition_plan : Callable, optional
             A plan that samples the beamline for some given inputs, by default default_acquisition_plan.
             Called directly in Agent, used only by __name__ in BlueskyAdaptiveAgent.
         digestion : Callable, optional
@@ -126,13 +109,15 @@ class BaseAgent:
             Whether to allow errors during acquistion. If `True`, errors will be caught as warnings, by default False
         sample_center_on_init : bool, optional
             Whether to sample the center of the DOF limits when the agent has no data yet, by default False
+        train_every : int, optional
+            How many times to train the model for each new point, by default 1
         """
         self.dofs = DOFList(list(dofs))
         self.objectives = ObjectiveList(list(objectives))
 
         _validate_dofs_and_objs(self.dofs, self.objectives)
 
-        self.acquisition_plan = acquistion_plan
+        self.acquisition_plan = acquisition_plan
         self.digestion = digestion
         self.digestion_kwargs = digestion_kwargs or {}
 
@@ -150,6 +135,7 @@ class BaseAgent:
         self.initialized = False
         self.a_priori_hypers = None
 
+        self.train_every = train_every
         self.n_last_trained = 0
 
     def raw_inputs(self, index: Optional[Union[str, int]] = None, **subset_kwargs) -> torch.Tensor:
@@ -159,6 +145,16 @@ class BaseAgent:
         if index is None:
             return torch.cat([self.raw_inputs(dof.name) for dof in self.dofs(**subset_kwargs)], dim=-1)
         return torch.tensor(self._table.loc[:, self.dofs[index].name].values, dtype=torch.double).unsqueeze(-1)
+
+    def argmax_best_f(self, weights: str = "default") -> int:
+        return int(self.scalarized_fitnesses(weights=weights, constrained=True).argmax())
+
+    @property
+    def random_ref_point(self) -> ArrayLike:
+        train_targets = self.train_targets(active=True, fitness=True, concatenate=True)
+        if not isinstance(train_targets, torch.Tensor):
+            raise RuntimeError("'random_ref_point' is not defined for multi-objective optimization.")
+        return train_targets[self.argmax_best_f(weights="random")]
 
     def train_inputs(self, index: Optional[Union[str, int]] = None, **subset_kwargs) -> torch.Tensor:
         """A two-dimensional tensor of all DOF values."""
@@ -207,14 +203,20 @@ class BaseAgent:
 
         return targets_dict
 
-    def _latent_dim_tuples(self, obj_index: Optional[Union[str, int]] = None) -> dict[str, list[tuple[int, ...]]]:
+    @property
+    def sample_domain(self) -> torch.Tensor:
+        """
+        Returns a (2, n_active_dof) array of lower and upper bounds for dofs.
+        Read-only DOFs are set to exactly their last known value.
+        Discrete DOFs are relaxed to some continuous domain.
+        """
+        return self.dofs(active=True).transform(self.dofs(active=True).search_domain.T).clone()
+
+    def _latent_dim_tuples(self, obj_index: Union[str, int]) -> list[tuple[int, ...]]:
         """
         For the objective indexed by 'obj_index', return a list of tuples, where each tuple represents
         a group of DOFs to fit a latent representation to.
         """
-
-        if obj_index is None:
-            return {obj.name: self._latent_dim_tuples(obj_index=obj.name) for obj in self.objectives}
 
         obj = self.objectives[obj_index]
 
@@ -227,6 +229,108 @@ class BaseAgent:
 
         u, uinv = np.unique(list(latent_group_index.values()), return_inverse=True)
         return [tuple(np.where(uinv == i)[0]) for i in range(len(u))]
+
+    @property
+    def evaluated_constraints(self) -> torch.Tensor:
+        constraint_objectives = self.objectives(constraint=True)
+        raw_targets_dict = self.raw_targets()
+        if len(constraint_objectives):
+            return torch.cat(
+                [obj.constrain(raw_targets_dict[obj.name]).unsqueeze(-1) for obj in constraint_objectives], dim=-1
+            )
+        else:
+            return torch.ones(size=(len(self._table), 0), dtype=torch.bool)
+
+    def scalarized_fitnesses(self, weights: str = "default", constrained: bool = True) -> torch.Tensor:
+        """
+        Return the scalar fitness for each sample, scalarized by the weighting scheme.
+
+        If constrained=True, the points that satisfy the most constraints are automatically better than the others.
+        """
+        fitness_objs = self.objectives(fitness=True)
+        if len(fitness_objs) >= 1:
+            f = self.fitness_scalarization(weights=weights).evaluate(
+                self.train_targets(active=True, fitness=True, concatenate=True)
+            )
+            f = torch.where(f.isnan(), -np.inf, f)  # remove all nans
+        else:
+            f = torch.zeros(len(self._table), dtype=torch.double)  # if there are no fitnesses, use a constant dummy fitness
+        if constrained:
+            # how many constraints are satisfied?
+            c = self.evaluated_constraints.sum(dim=-1)
+            f = torch.where(c < c.max(), -np.inf, f)
+        return f
+
+    def best_f(self, weights: str = "default") -> float:
+        return float(self.scalarized_fitnesses(weights=weights, constrained=True).max())
+
+    def fitness_scalarization(self, weights: Union[str, torch.Tensor] = "default") -> ScalarizedPosteriorTransform:
+        active_fitness_objectives = self.objectives(active=True, fitness=True)
+        if len(active_fitness_objectives) == 0:
+            return ScalarizedPosteriorTransform(weights=torch.tensor([1.0], dtype=torch.double))
+        if weights == "default":
+            weights = torch.tensor([obj.weight for obj in active_fitness_objectives], dtype=torch.double)
+        elif weights == "equal":
+            weights = torch.ones(len(active_fitness_objectives), dtype=torch.double)
+        elif weights == "random":
+            weights = torch.rand(len(active_fitness_objectives), dtype=torch.double)
+            weights *= len(active_fitness_objectives) / weights.sum()
+        elif not isinstance(weights, torch.Tensor):
+            raise ValueError(f"'weights' must be a Tensor or one of ['default', 'equal', 'random'], and not {weights}.")
+        return ScalarizedPosteriorTransform(weights=weights)
+
+    @property
+    def fitness_model(self) -> Model:
+        active_fitness_objectives = self.objectives(active=True, fitness=True)
+        if len(active_fitness_objectives) == 0:
+            # A dummy model that outputs noise, for when there are only constraints.
+            dummy_X = self.sample(n=256, normalize=True).squeeze(-2)
+            dummy_Y = torch.rand(size=(*dummy_X.shape[:-1], 1), dtype=torch.double)
+            return construct_single_task_model(X=dummy_X, y=dummy_Y, min_noise=1e2, max_noise=2e2)
+        if len(active_fitness_objectives) == 1:
+            return active_fitness_objectives[0].model
+        return ModelListGP(*[obj.model for obj in active_fitness_objectives])
+
+    def constraint(self, x: torch.Tensor) -> torch.Tensor:
+        log_p = torch.zeros(x.shape[:-1])
+        for obj in self.objectives(active=True):
+            log_p += obj.log_total_constraint(x)
+
+        return log_p.exp()  # + 1e-6 * normalize(x, self.sample_domain).square().sum(axis=-1)
+
+    def sample(self, n: int = DEFAULT_MAX_SAMPLES, normalize: bool = False, method: str = "quasi-random") -> torch.Tensor:
+        """
+        Returns a (..., 1, n_active_dofs) tensor of points sampled within the parameter space.
+
+        Parameters
+        ----------
+        n : int
+            How many points to sample.
+        method : str
+            How to sample the points. Must be one of 'quasi-random', 'random', or 'grid'.
+        normalize: bool
+            If True, sample the unit hypercube. If False, sample the parameter space of the agent.
+        """
+
+        active_dofs = self.dofs(active=True)
+
+        if method == "quasi-random":
+            X = utils.normalized_sobol_sampler(n, d=len(active_dofs))
+
+        elif method == "random":
+            X = torch.rand(size=(n, 1, len(active_dofs)))
+
+        elif method == "grid":
+            n_side_if_settable = int(np.power(n, 1 / np.sum(~active_dofs.read_only)))
+            sides = [
+                torch.linspace(0, 1, n_side_if_settable) if not dof.read_only else torch.zeros(1) for dof in active_dofs
+            ]
+            X = torch.cat([x.unsqueeze(-1) for x in torch.meshgrid(sides, indexing="ij")], dim=-1).unsqueeze(-2).double()
+
+        else:
+            raise ValueError("'method' argument must be one of ['quasi-random', 'random', 'grid'].")
+
+        return X.double() if normalize else self.dofs(active=True).untransform(X).double()
 
     # @property
     def pruned_mask(self) -> torch.Tensor:
@@ -249,9 +353,9 @@ class BaseAgent:
         Read-only: constrain to the readback value
         """
 
-        return Normalize(d=self.dofs.active.sum())
+        return Normalize(d=len(self.dofs(active=True)))
 
-    def _construct_model(self, obj, skew_dims: Optional[list[tuple[int, ...]]] = None):
+    def _construct_model(self, obj, skew_dims: Optional[list[tuple[int, ...]]] = None) -> None:
         """
         Construct an untrained model for an objective.
         """
@@ -261,8 +365,8 @@ class BaseAgent:
         train_inputs = self.train_inputs(active=True)
         train_targets = self.train_targets()[obj.name].unsqueeze(-1)
 
-        inputs_are_trusted = ~torch.isnan(train_inputs).any(axis=1)
-        targets_are_trusted = ~torch.isnan(train_targets).any(axis=1)
+        inputs_are_trusted = ~torch.isnan(train_inputs).any(dim=1)
+        targets_are_trusted = ~torch.isnan(train_targets).any(dim=1)
 
         trusted = inputs_are_trusted & targets_are_trusted & ~self.pruned_mask()
 
@@ -271,7 +375,7 @@ class BaseAgent:
             y=train_targets[trusted],
             min_noise=obj.min_noise,
             max_noise=obj.max_noise,
-            skew_dims=self._latent_dim_tuples()[obj.name],
+            skew_dims=self._latent_dim_tuples(obj.name),
         )
 
         obj.model_dofs = set(self.dofs(active=True).names)  # if these change, retrain the model on self.ask()
@@ -300,7 +404,7 @@ class BaseAgent:
     def update_models(
         self,
         train: Optional[bool] = None,
-    ):
+    ) -> None:
         objectives_to_model = self.objectives if self.model_inactive_objectives else self.objectives(active=True)
         for obj in objectives_to_model:
             t0 = ttime.monotonic()
@@ -332,7 +436,7 @@ class BaseAgent:
         append: bool = True,
         update_models: bool = True,
         train: Optional[bool] = None,
-    ):
+    ) -> None:
         """
         Inform the agent about new inputs and targets for the model.
 
@@ -353,9 +457,12 @@ class BaseAgent:
         """
 
         if not data:
-            if not x and y:
+            if x and y and metadata:
+                data = {**x, **y, **metadata}
+            elif x and y:
+                data = {**x, **y}
+            else:
                 raise ValueError("Must supply either x and y, or data.")
-            data = {**x, **y, **metadata}
 
         data = {k: list(np.atleast_1d(v)) for k, v in data.items()}
         unique_field_lengths = {len(v) for v in data.values()}
@@ -366,7 +473,7 @@ class BaseAgent:
         # TODO: This is an innefficient approach to caching data. Keep a list, make table at update model time.
         new_table = pd.DataFrame(data)
         self._table = pd.concat([self._table, new_table]) if append else new_table
-        self._table.index = np.arange(len(self._table))
+        self._table.index = pd.Index(np.arange(len(self._table)))
         if update_models:
             self.update_models(train=train)
 
@@ -477,199 +584,6 @@ class BaseAgent:
         return res
 
 
-class BlueskyAdaptiveAgent(BaseAgent, BlueskyAdaptiveBaseAgent):
-    def __init__(
-        self,
-        *,
-        acqf_string: str,
-        route: bool,
-        sequential: bool,
-        upsample: int,
-        acqf_kwargs: dict[str, Any],
-        detector_names: Optional[list[str]] = (),
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-        self._acqf_string = acqf_string
-        self._route = route
-        self._sequential = sequential
-        self._upsample = upsample
-        self._acqf_kwargs = acqf_kwargs
-        self._detector_names = detector_names or []
-
-    @property
-    def detector_names(self) -> list[str]:
-        return [str(name) for name in self._detector_names]
-
-    @detector_names.setter
-    def detector_names(self, names: list[str]):
-        self._detector_names = list(names)
-
-    @property
-    def acquisition_function(self) -> str:
-        return str(self._acqf_string)
-
-    @acquisition_function.setter
-    def acquisition_function(self, acqf_string: str):
-        self._acqf_string = str(acqf_string)
-
-    @property
-    def route(self) -> bool:
-        return bool(self._route)
-
-    @route.setter
-    def route(self, route: bool):
-        self._route = route
-
-    @property
-    def sequential(self) -> bool:
-        return bool(self._sequential)
-
-    @sequential.setter
-    def sequential(self, sequential: bool):
-        self._sequential = sequential
-
-    @property
-    def upsample(self) -> int:
-        return int(self._upsample)
-
-    @upsample.setter
-    def upsample(self, upsample: int):
-        self._upsample = int(upsample)
-
-    @property
-    def acqf_kwargs(self) -> dict[str, str]:
-        return {str(k): str(v) for k, v in self._acqf_kwargs.items()}
-
-    def update_acqf_kwargs(self, **kwargs):
-        self._acqf_kwargs.update(kwargs)
-
-    @property
-    def detector_names(self) -> list[str]:
-        return [str(name) for name in self._detector_names]
-
-    @detector_names.setter
-    def detector_names(self, names: list[str]):
-        self._detector_names = list(names)
-
-    def server_registrations(self) -> list[str]:
-        """This is how we make these avaialble to the REST API."""
-        self._register_method("Update Acquistion Function Kwargs", self.update_acqf_kwargs)
-        self._register_property("Acquisition Function", self.acquisition_function, self.acquisition_function)
-        self._register_property("Route Points", self.route, self.route)
-        self._register_property("Sequential Points", self.sequential, self.sequential)
-        self._register_property("Upsample Points", self.upsample, self.upsample)
-        return super().server_registrations()
-
-    def ask(self, batch_size: int) -> tuple[Sequence[dict[str, ArrayLike]], Sequence[ArrayLike]]:
-        default_result = super().ask(
-            n=batch_size,
-            acqf=self._acqf_string,
-            route=self._route,
-            sequential=self._sequential,
-            upsample=self._upsample,
-            **self._acqf_kwargs,
-        )
-
-        """res = {
-            "points": {dof.name: list(points[..., i]) for i, dof in enumerate(active_dofs(read_only=False))},
-            "acqf_name": acqf_config["name"],
-            "acqf_obj": list(np.atleast_1d(acqf_obj.numpy())),
-            "acqf_kwargs": acqf_kwargs,
-            "duration_ms": duration,
-            "sequential": sequential,
-            "upsample": upsample,
-            "read_only_values": read_only_values,
-            # "posterior": p,
-        }
-        """
-
-        points: dict[str, list[ArrayLike]] = default_result.pop("points")
-        acqf_obj: list[ArrayLike] = default_result.pop("acqf_obj")
-        # Turn dict of list of points into list of consistently sized points
-        points: list[tuple[ArrayLike]] = list(zip(*[value for _, value in points.items()]))
-        dicts = []
-        for point, obj in zip(points, acqf_obj):
-            d = default_result.copy()
-            d["point"] = point
-            d["acqf_obj"] = obj
-            dicts.append(d)
-        return points, dicts
-
-    def tell(self, x: dict[str, ArrayLike], y: dict[str, ArrayLike]):
-        x = {key: x_i for x_i, key in zip(x, self.dofs.names)}
-        y = {key: y_i for y_i, key in zip(y, self.objectives.names)}
-        super().tell(data={**x, **y})
-        return {**x, **y}
-
-    def report(self) -> dict[str, Any]:
-        raise NotImplementedError("Report is not implmented for BlueskyAdaptiveAgent")
-
-    def unpack_run(self, run: BlueskyRun) -> tuple[list[ArrayLike], list[ArrayLike]]:
-        """Use my DOFs to convert the run into an independent array, and my objectives to create the dependent array.
-        In practice for shape management, we will use lists not np.arrays at this stage.
-        Parameters
-        ----------
-        run : BlueskyRun
-
-        Returns
-        -------
-        independent_var :
-            The independent variable of the measurement
-        dependent_var :
-            The measured data, processed for relevance
-        """
-        if not self.digestion or self.digestion == default_digestion_function:
-            # Assume all raw data is available in primary stream as keys
-            return (
-                [run.primary.data[key].read() for key in self.dofs.names],
-                [run.primary.data[key].read() for key in self.objectives.names],
-            )
-        else:
-            # Hope and pray that the digestion function designed for DataFrame can handle the XArray
-            data: pd.DataFrame = self.digestion(run.primary.data.read(), **self.digestion_kwargs)
-            return [data.loc[:, key] for key in self.dofs.names], [data.loc[:, key] for key in self.objectives.names]
-
-    def measurement_plan(self, point: ArrayLike) -> tuple[str, list[Any], dict[str, Any]]:
-        """Fetch the string name of a registered plan, as well as the positional and keyword
-        arguments to pass that plan.
-
-        Args/Kwargs is a common place to transform relative into absolute motor coords, or
-        other device specific parameters.
-
-        By default, this measurement plan attempts to use in the built in functionality in a QueueServer compatible way.
-        Signals and Devices are not passed as objects, but serialized as strings for the RE as a service to use.
-
-        Parameters
-        ----------
-        point : ArrayLike
-            Next point to measure using a given plan
-
-        Returns
-        -------
-        plan_name : str
-        plan_args : List
-            List of arguments to pass to plan from a point to measure.
-        plan_kwargs : dict
-            Dictionary of keyword arguments to pass the plan, from a point to measure.
-        """
-        if isinstance(self.acquisition_plan, Callable):
-            plan_name = self.acquisition_plan.__name__
-        else:
-            plan_name = self.acquisition_plan
-        if plan_name == "default_acquisition_plan":
-            # Convert point back to dict form for the sake of compatability with default plan
-            acquisition_dofs = self.dofs(active=True, read_only=False)
-
-            return self.acquisition_plan.__name__, [
-                acquisition_dofs,
-                {dof.name: point[i] for i, dof in enumerate(acquisition_dofs)},
-                [*self.detector_names, *[dev.__name__ for dev in self.dofs.devices]],
-            ]
-        else:
-            raise NotImplementedError("Only default_acquisition_plan is implemented")
-
-
 class Agent(BaseAgent):
     def __init__(
         self,
@@ -677,7 +591,7 @@ class Agent(BaseAgent):
         objectives: Sequence[Objective],
         db: Optional[Broker] = None,
         detectors: Optional[Sequence[Signal]] = None,
-        acquistion_plan: Callable = default_acquisition_plan,
+        acquisition_plan: Callable = default_acquisition_plan,
         digestion: Callable = default_digestion_function,
         digestion_kwargs: Optional[dict] = None,
         verbose: bool = False,
@@ -687,7 +601,6 @@ class Agent(BaseAgent):
         tolerate_acquisition_errors: bool = False,
         sample_center_on_init: bool = False,
         trigger_delay: float = 0,
-        train_every: int = 1,
     ):
         """
         A Bayesian optimization agent.
@@ -739,7 +652,7 @@ class Agent(BaseAgent):
         super().__init__(
             dofs=dofs,
             objectives=objectives,
-            acquistion_plan=acquistion_plan,
+            acquisition_plan=acquisition_plan,
             digestion=digestion,
             digestion_kwargs=digestion_kwargs,
             verbose=verbose,
@@ -754,7 +667,6 @@ class Agent(BaseAgent):
 
         self.db = db
 
-        self.train_every = train_every
         self.trigger_delay = trigger_delay
 
         self.initialized = False
@@ -784,40 +696,6 @@ class Agent(BaseAgent):
     def redigest(self):
         self._table = self.digestion(self._table, **self.digestion_kwargs)
 
-    def sample(self, n: int = DEFAULT_MAX_SAMPLES, normalize: bool = False, method: str = "quasi-random") -> torch.Tensor:
-        """
-        Returns a (..., 1, n_active_dofs) tensor of points sampled within the parameter space.
-
-        Parameters
-        ----------
-        n : int
-            How many points to sample.
-        method : str
-            How to sample the points. Must be one of 'quasi-random', 'random', or 'grid'.
-        normalize: bool
-            If True, sample the unit hypercube. If False, sample the parameter space of the agent.
-        """
-
-        active_dofs = self.dofs(active=True)
-
-        if method == "quasi-random":
-            X = utils.normalized_sobol_sampler(n, d=len(active_dofs))
-
-        elif method == "random":
-            X = torch.rand(size=(n, 1, len(active_dofs)))
-
-        elif method == "grid":
-            n_side_if_settable = int(np.power(n, 1 / np.sum(~active_dofs.read_only)))
-            sides = [
-                torch.linspace(0, 1, n_side_if_settable) if not dof.read_only else torch.zeros(1) for dof in active_dofs
-            ]
-            X = torch.cat([x.unsqueeze(-1) for x in torch.meshgrid(sides, indexing="ij")], dim=-1).unsqueeze(-2).double()
-
-        else:
-            raise ValueError("'method' argument must be one of ['quasi-random', 'random', 'grid'].")
-
-        return X.double() if normalize else self.dofs(active=True).untransform(X).double()
-
     def learn(
         self,
         acqf: str = "qei",
@@ -829,7 +707,7 @@ class Agent(BaseAgent):
         hypers: Optional[str] = None,
         route: bool = True,
         **acqf_kwargs,
-    ):
+    ) -> Generator[Msg, None, None]:
         """This returns a Bluesky plan which iterates the learning algorithm, looping over ask -> acquire -> tell.
 
         For example:
@@ -886,7 +764,7 @@ class Agent(BaseAgent):
             The thing to be viewed. Either 'mean', 'error', or an acquisition function.
         """
 
-        import napari  # noqa E402
+        import napari  # type: ignore
 
         test_grid = self.sample(n=max_inputs, method="grid")
 
@@ -917,7 +795,7 @@ class Agent(BaseAgent):
 
         self.viewer.dims.axis_labels = self.dofs.names
 
-    def acquire(self, points: ArrayLike) -> pd.DataFrame:
+    def acquire(self, points: dict[str, list[ArrayLike]]) -> Generator[Msg, None, pd.DataFrame]:
         """Acquire and digest according to the self's acquisition and digestion plans.
 
         Parameters
@@ -1017,18 +895,6 @@ class Agent(BaseAgent):
         """A model encompassing all the objectives. A single GP in the single-objective case, or a model list."""
         return self.model.posterior(self.dofs(active=True).transform(torch.tensor(x)))
 
-    @property
-    def fitness_model(self) -> Model:
-        active_fitness_objectives = self.objectives(active=True, fitness=True)
-        if len(active_fitness_objectives) == 0:
-            # A dummy model that outputs noise, for when there are only constraints.
-            dummy_X = self.sample(n=256, normalize=True).squeeze(-2)
-            dummy_Y = torch.rand(size=(*dummy_X.shape[:-1], 1), dtype=torch.double)
-            return construct_single_task_model(X=dummy_X, y=dummy_Y, min_noise=1e2, max_noise=2e2)
-        if len(active_fitness_objectives) == 1:
-            return active_fitness_objectives[0].model
-        return ModelListGP(*[obj.model for obj in active_fitness_objectives])
-
     # @property
     # def pseudofitness_model(self):
     #     """
@@ -1044,58 +910,6 @@ class Agent(BaseAgent):
     #     if len(active_fitness_objectives) == 1:
     #         return active_fitness_objectives[0].model
     #     return ModelListGP(*[obj.model for obj in active_fitness_objectives])
-
-    @property
-    def evaluated_constraints(self) -> torch.Tensor:
-        constraint_objectives = self.objectives(constraint=True)
-        raw_targets_dict = self.raw_targets()
-        if len(constraint_objectives):
-            return torch.cat(
-                [obj.constrain(raw_targets_dict[obj.name]).unsqueeze(-1) for obj in constraint_objectives], dim=-1
-            )
-        else:
-            return torch.ones(size=(len(self._table), 0), dtype=torch.bool)
-
-    def fitness_scalarization(self, weights: str = "default") -> ScalarizedPosteriorTransform:
-        active_fitness_objectives = self.objectives(active=True, fitness=True)
-        if len(active_fitness_objectives) == 0:
-            return ScalarizedPosteriorTransform(weights=torch.tensor([1.0], dtype=torch.double))
-        if weights == "default":
-            weights = torch.tensor([obj.weight for obj in active_fitness_objectives], dtype=torch.double)
-        elif weights == "equal":
-            weights = torch.ones(len(active_fitness_objectives), dtype=torch.double)
-        elif weights == "random":
-            weights = torch.rand(len(active_fitness_objectives), dtype=torch.double)
-            weights *= len(active_fitness_objectives) / weights.sum()
-        elif not isinstance(weights, torch.Tensor):
-            raise ValueError(f"'weights' must be a Tensor or one of ['default', 'equal', 'random'], and not {weights}.")
-        return ScalarizedPosteriorTransform(weights=weights)
-
-    def scalarized_fitnesses(self, weights: str = "default", constrained: bool = True) -> torch.Tensor:
-        """
-        Return the scalar fitness for each sample, scalarized by the weighting scheme.
-
-        If constrained=True, the points that satisfy the most constraints are automatically better than the others.
-        """
-        fitness_objs = self.objectives(fitness=True)
-        if len(fitness_objs) >= 1:
-            f = self.fitness_scalarization(weights=weights).evaluate(
-                self.train_targets(active=True, fitness=True, concatenate=True)
-            )
-            f = torch.where(f.isnan(), -np.inf, f)  # remove all nans
-        else:
-            f = torch.zeros(len(self._table), dtype=torch.double)  # if there are no fitnesses, use a constant dummy fitness
-        if constrained:
-            # how many constraints are satisfied?
-            c = self.evaluated_constraints.sum(axis=-1)
-            f = torch.where(c < c.max(), -np.inf, f)
-        return f
-
-    def argmax_best_f(self, weights: str = "default") -> int:
-        return int(self.scalarized_fitnesses(weights=weights, constrained=True).argmax())
-
-    def best_f(self, weights: str = "default") -> float:
-        return float(self.scalarized_fitnesses(weights=weights, constrained=True).max())
 
     @property
     def pareto_mask(self) -> torch.Tensor:
@@ -1123,10 +937,6 @@ class Agent(BaseAgent):
     def min_ref_point(self) -> ArrayLike:
         y = self.train_targets(concatenate=True)[:, self.objectives.type == "fitness"]
         return y[y.argmax(axis=0)].min(axis=0).values
-
-    @property
-    def random_ref_point(self) -> ArrayLike:
-        return self.train_targets(active=True, fitness=True, concatenate=True)[self.argmax_best_f(weights="random")]
 
     @property
     def all_objectives_valid(self) -> torch.Tensor:
@@ -1159,15 +969,6 @@ class Agent(BaseAgent):
         """
 
         return acquisition._construct_acqf(self, identifier=identifier, return_metadata=return_metadata)
-
-    @property
-    def sample_domain(self) -> torch.Tensor:
-        """
-        Returns a (2, n_active_dof) array of lower and upper bounds for dofs.
-        Read-only DOFs are set to exactly their last known value.
-        Discrete DOFs are relaxed to some continuous domain.
-        """
-        return self.dofs(active=True).transform(self.dofs(active=True).search_domain.T).clone()
 
     def save_data(self, path: str = "./data.h5"):
         """
@@ -1209,13 +1010,6 @@ class Agent(BaseAgent):
         for obj in self.objectives(active=True):
             obj.model.load_state_dict(hypers[obj.name])
         self.validity_constraint.load_state_dict(hypers["validity_constraint"])
-
-    def constraint(self, x: ArrayLike) -> torch.Tensor:
-        log_p = torch.zeros(x.shape[:-1])
-        for obj in self.objectives(active=True):
-            log_p += obj.log_total_constraint(x)
-
-        return log_p.exp()  # + 1e-6 * normalize(x, self.sample_domain).square().sum(axis=-1)
 
     @property
     def hypers(self) -> dict[str, dict[str, dict[str, torch.Tensor]]]:
